Fix image and script paths in tutorial
Context: the introduction to R tutorial has been moved
into the educer package

Problem: the tutorial doesn't run yet bc the figures
and the example scripts need to be added to the resources
folder and all the paths needs to be fixed

Solution: add the figures and fix the paths. Use the
setup_resources() function to setup the paths.

Remaining problems: when I run the tutorial it says that
it cannot find the figures.