Remove "How to follow tutorial" & "Data in R"
In this commit I removed a couple of sections and
modified the tutorial accordingly.

1/ The section on how to follow the tutorial will be
moved to its own child document (currently living
inside resources/scripts but eventually going to
live in resources/markdown. The child document has
not yet been added because I still couldn't figure
it out.

2/ The section on Data in R was mostly base R. There
was some stuff about vectors vs data frames and on
data types that is covered in the intro to R
tutorial. There was a lot about indexing which is
not otherwise needed. There was also a bit about
logical operators that was only needed for `filter`
so I now only introduce the logical operators in
the `filter` section.