Introduce child documents
Created two Rmd files that like in:
inst/resources/markdown - one is a short data
description, the other is a short "how to follow
this tutorial".

Replaced or modified the data description and how-to
sections of both Tidyverse and Intro R and Rstudio
tutorials; they now use the child documents.